connectors/msdefender2tone/pyproject.toml
[project]
name = "msdefender2tone"
version = "1.0.0"
description = "MS Defender to Tenable One"
authors = [
    {name = "Tenable, Inc", email = "ecosystem-support@tenable.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "restfly>=1.5.0",
    "rich>=13.9.3",
    "tomlkit>=0.13.2",
    "typer>=0.12.5",
    "pytenable>=1.7.1",
    "pydantic-extra-types>=2.9.0",
    "tenint>=0.1.3",
]

[project.optional-dependencies]
testing = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "responses>=0.25.3",
]

[project.urls]
repository = "https://github.com/tenable/tenable-connectors"
logo = "https://raw.githubusercontent.com/tenable/tenable-connectors/refs/heads/main/connectors/msdefender2tone/logo.svg"
support = "https://community.tenable.com"

[tool.tenint.connector]
title = "Microsoft Defender to Tenable One"
tags = ["msdefender", "tvm"]


[tool.ruff]
line-length = 88
indent-width = 4
exclude = [
  ".nova",
  ".github",
  ".git",
  ".pytest_cache",
  "__pycache__"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
fixable = [ "ALL" ]
unfixable = [ "B" ]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--cov=msdefender --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]

